app/pyproject.toml
[tool.poetry]
name = "tronql"
version = "0.1.0"
description = ""
authors = ["FerumFlex <ferumflex@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = {extras = ["standard"], version = "^0.18.3"}
fastapi = "^0.85.0"
strawberry-graphql = {extras = ["debug-server"], version = "^0.133.5"}
SQLAlchemy = "^1.4.41"
asyncpg = "^0.26.0"
psycopg2-binary = "^2.9.4"
alembic = "^1.8.1"
pydantic = "^1.10.2"
sentry-sdk = "^1.9.10"
faust-streaming = "^0.8.11"
aiohttp = "^3.8.3"
base58 = "^2.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
